fix: test

+++ src/lane_detection/main.cpp
@@ -291,7 +291,7 @@ int main()
                         average_angle /= 5;
                     }
                     
-                    int speed = int(50 * (100 / (std::abs(average_angle) + 100)));
+                    int speed = int(50 * float(100.0f / float(std::abs(float(average_angle)) + 100.0f)));
 
                     logger->log("%d, %d", speed, average_angle);
